update mastermind
implemented tipEqualCode() as method that finds out if user found the
code, if so terminates game.
rearranged main method, cut code into pieces and filled it in methods.
